+++ src/components/common/Button.jsx
@@ -1,6 +1,6 @@
 import { memo } from 'react';
 import PropTypes from 'prop-types';
-import styles from 'css/Button.module.css';
+import styles from 'css/module/Button.module.css';
 
 const Button = ({ className, theme, children }) => {
   return <button className={`${styles.container} ${className} font-${theme}`}>{children}</button>;
