+++ src/components/common/Header.jsx
@@ -1,4 +1,4 @@
-import styles from 'css/Header.module.css';
+import styles from 'css/module/Header.module.css';
 import useHeader from 'hooks/useHeader';
 
 const Header = () => {
