+++ src/css/input.css
@@ -26,15 +26,3 @@
   font-size: 12px;
   color: #525252;
 }
-
-.input-underline {
-  text-align: center;
-  border: none;
-  background: none;
-  outline: none;
-
-  margin: 33px 0;
-  padding: 4px 0;
-
-  border-bottom: 1px solid #383838;
-}
