+++ src/css/module/ConfirmationPage.module.css
@@ -13,3 +13,15 @@
   color: #383838;
   margin-bottom: 77px;
 }
+
+.input {
+  text-align: center;
+  border: none;
+  background: none;
+  outline: none;
+
+  margin: 33px 0;
+  padding: 4px 0;
+
+  border-bottom: 1px solid #383838;
+}
