+++ src/css/module/MainPage.module.css
@@ -1,4 +1,4 @@
-.main-container {
+.container {
   overflow: scroll;
   height: 80%;
   display: flex;
@@ -6,17 +6,17 @@
   align-items: center;
 }
 
-.card-list {
+.cardList {
   display: flex;
   flex-direction: column;
   align-items: center;
 }
 
-.alias {
-  text-align: center;
-}
-
-.card-add-icon {
+.addIcon {
   cursor: pointer;
   margin: 30px 0;
+}
+
+.alias {
+  text-align: center;
 }
